Update the getClientIP() function in Request.php
Update the getClientIP() function to properly return the last IP address in the X-Forwarded-For header

The syntax for this header is a comma separated list of proxied IP address

X-Forwarded-For: <client>, <proxy1>, <proxy2>

See: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For

As is the getClientIp function is returning the raw value of X-Forwarded-For which if proxied can contain multiple IP address.

This causes issues if this function is used to get the client's IP address - e.g. in payment extension to validate the client's IP Address.